2-backend/requirements.txt
# Circuit Simulator Backend - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSocket Support
websockets==12.0
python-socketio==5.10.0

# Validation
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Scientific Computing (for simulation)
numpy==1.26.2
scipy==1.11.4

# Advanced Circuit Simulation
PySpice==1.5
ngspice==0.1.0

# Component Database & Pricing
octopart==0.0.7
requests==2.31.0

# Digital Logic Simulation
bitstring==4.1.4

# BOM & Data Management
pandas==2.1.4
openpyxl==3.1.2

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# Utilities
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
